server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.121.2",
    "google-adk>=1.18.0",
    "langchain-community>=0.4.1",
    "pyyaml>=6.0.3",
]
